Fix yi-34b tokenizer, use_fast=False (#241)
* fix : use_fast=not args.disable_fast_tokenizer

* "--disable_fast_tokenizer "

* fix

+++ openrlhf/datasets/reward_dataset.py
@@ -125,7 +125,7 @@ class RewardDataset(Dataset):
                     truncation=True,
                     return_tensors="pt",
                 )
-                prompt_ids_len = prompt_token["attention_mask"].sum().item()
+                prompt_ids_len = prompt_token["attention_mask"].int().sum().item()
                 # filter the sample whose length is greater than max_length (2 for answer length)
                 if prompt_ids_len >= self.max_length - 2:
                     continue
